+++ scripts/settings.js
@@ -1,20 +1,31 @@
 import { MODULE_NAME } from "./const.js";
 
+export const SETTING_NAMES = {
+    AUTO_CHECK: "autoCheck",
+    AUTO_DMG: "autoDamage",
+    SHOW_TOTAL_DMG: "showTotalDamage",
+    ROLL_DLG_BHVR: "rollDialogBehavior",
+    SHOW_ROLL_DLG_MOD: "showRollDialogModifier",
+    ADV_MOD: "advModifier",
+    DISADV_MOD: "disAdvModifier",
+}
+
 export function registerSettings() {
     _registerAutoRollsSettings();
     _registerOtherSettings();
     _registerModifierKeySettings();
 }
 
-export function getModifierSettingLocalOrDefault(settingKey) {
+export function getSettingLocalOrDefault(settingKey) {
     const localKey = `${settingKey}Local`;
     const globalKey = `${settingKey}Global`;
     const value = game.settings.get(MODULE_NAME, localKey);
-    return value === "default" ? game.settings.get(MODULE_NAME, globalKey) : value;
+    const defaultValue = game.settings.settings.get(`${MODULE_NAME}.${localKey}`).default;
+    return value === defaultValue ? game.settings.get(MODULE_NAME, globalKey) : value;
 }
 
 function _registerAutoRollsSettings() {
-    game.settings.register(MODULE_NAME, "autoCheck", {
+    game.settings.register(MODULE_NAME, SETTING_NAMES.AUTO_CHECK, {
         name: game.i18n.localize(`${MODULE_NAME}.SETTINGS.AutoRollChecksLabel`),
         hint: game.i18n.localize(`${MODULE_NAME}.SETTINGS.AutoRollChecksHint`),
         scope: "world",
@@ -23,7 +34,7 @@ function _registerAutoRollsSettings() {
         default: false,
     });
 
-    game.settings.register(MODULE_NAME, "autoDamage", {
+    game.settings.register(MODULE_NAME, SETTING_NAMES.AUTO_DMG, {
         name: game.i18n.localize(`${MODULE_NAME}.SETTINGS.AutoRollDamageLabel`),
         hint: game.i18n.localize(`${MODULE_NAME}.SETTINGS.AutoRollDamageHint`),
         scope: "world",
@@ -34,7 +45,7 @@ function _registerAutoRollsSettings() {
 }
 
 function _registerOtherSettings() {
-    game.settings.register(MODULE_NAME, "showTotalDamage", {
+    game.settings.register(MODULE_NAME, SETTING_NAMES.SHOW_TOTAL_DMG, {
         name: game.i18n.localize(`${MODULE_NAME}.SETTINGS.ShowDamageTotalLabel`),
         hint: game.i18n.localize(`${MODULE_NAME}.SETTINGS.ShowDamageTotalHint`),
         scope: "world",
@@ -42,22 +53,33 @@ function _registerOtherSettings() {
         type: Boolean,
         default: false,
     });
+
+    const globalEnabledChoices = {
+        "skip": game.i18n.localize(`${MODULE_NAME}.SETTINGS.Skip`),
+        "show": game.i18n.localize(`${MODULE_NAME}.SETTINGS.Show`),
+    };
+
+    const localEnabledChoices = {
+        "default": game.i18n.localize(`${MODULE_NAME}.MODIFIERS.default`),
+        ...globalEnabledChoices,
+    };
+
+    const _registerToggle = makeGlobalLocalSettingRegisterer(globalEnabledChoices, localEnabledChoices);
+
+    const rollDialogBehaviorLabel = game.i18n.localize(`${MODULE_NAME}.SETTINGS.RollDialogBehaviorLabel`);
+    const rollDialogBehaviorHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.RollDialogBehaviorHint`);
+
+    _registerToggle(`${SETTING_NAMES.ROLL_DLG_BHVR}Global`, rollDialogBehaviorLabel, rollDialogBehaviorHint, true, "skip");
+    _registerToggle(`${SETTING_NAMES.ROLL_DLG_BHVR}Local`, rollDialogBehaviorLabel, rollDialogBehaviorHint, false, "default");
 }
 
 function _registerModifierKeySettings() {
-    const globalDefault = game.i18n.localize(`${MODULE_NAME}.SETTINGS.GlobalDefault`);
-    const localOverride = game.i18n.localize(`${MODULE_NAME}.SETTINGS.LocalOverride`);
-
-    const globalHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.GlobalHint`);
-    const localHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.LocalHint`);
-
-    const showRollDialogLabel = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierShowRollDialogLabel`);
-    const showRollDialogHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierShowRollDialogHint`);
+    const showRollDialogModifierLabel = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierShowRollDialogLabel`);
+    const showRollDialogModifierHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierShowRollDialogHint`);
     const advModifierLabel = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierAdvLabel`);
     const advModifierHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierAdvHint`);
     const disAdvModifierLabel = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierDisAdvLabel`);
     const disAdvModifierHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.ModifierDisAdvHint`);
-
 
     const globalModifierChoices = {
         "altKey": game.i18n.localize(`${MODULE_NAME}.MODIFIERS.altKey`),
@@ -66,16 +88,30 @@ function _registerModifierKeySettings() {
         "shiftKey": game.i18n.localize(`${MODULE_NAME}.MODIFIERS.shiftKey`),
     };
 
-    const localModifierChoices = mergeObject(
-        { "default": game.i18n.localize(`${MODULE_NAME}.MODIFIERS.default`) },
-        globalModifierChoices,
-        { inplace: false },
-    );
+    const localModifierChoices = {
+        "default": game.i18n.localize(`${MODULE_NAME}.MODIFIERS.default`),
+        ...globalModifierChoices,
+    }
 
-    const makeName = (name, modifier) => `${name} (${modifier})`;
-    const makeHint = (hint, modifier) => `${hint} ${modifier}`;
+    const _registerModifier = makeGlobalLocalSettingRegisterer(globalModifierChoices, localModifierChoices);
 
-    function _registerModifier(key, label, hint, isGlobal, defaultValue) {
+    _registerModifier(`${SETTING_NAMES.SHOW_ROLL_DLG_MOD}Global`, showRollDialogModifierLabel, showRollDialogModifierHint, true, "shiftKey");
+    _registerModifier(`${SETTING_NAMES.ADV_MOD}Global`, advModifierLabel, advModifierHint, true, "altKey");
+    _registerModifier(`${SETTING_NAMES.DISADV_MOD}Global`, disAdvModifierLabel, disAdvModifierHint, true, "ctrlKey");
+
+    _registerModifier(`${SETTING_NAMES.SHOW_ROLL_DLG_MOD}Local`, showRollDialogModifierLabel, showRollDialogModifierHint, false, "default");
+    _registerModifier(`${SETTING_NAMES.ADV_MOD}Local`, advModifierLabel, advModifierHint, false, "default");
+    _registerModifier(`${SETTING_NAMES.DISADV_MOD}Local`, disAdvModifierLabel, disAdvModifierHint, false, "default");
+}
+
+function makeGlobalLocalSettingRegisterer(globalChoices, localChoices) {
+    const globalDefault = game.i18n.localize(`${MODULE_NAME}.SETTINGS.GlobalDefault`);
+    const localOverride = game.i18n.localize(`${MODULE_NAME}.SETTINGS.LocalOverride`);
+
+    const globalHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.GlobalHint`);
+    const localHint = game.i18n.localize(`${MODULE_NAME}.SETTINGS.LocalHint`);
+
+    return function registerer(key, label, hint, isGlobal, defaultValue) {
         game.settings.register(MODULE_NAME, key, {
             name: makeName(label, isGlobal ? globalDefault : localOverride),
             hint: makeHint(hint, isGlobal ? globalHint : localHint),
@@ -83,15 +119,10 @@ function _registerModifierKeySettings() {
             config: true,
             type: String,
             default: defaultValue,
-            choices: isGlobal ? globalModifierChoices : localModifierChoices,
+            choices: isGlobal ? globalChoices : localChoices,
         });
     }
-
-    _registerModifier("showRollDialogModifierGlobal", showRollDialogLabel, showRollDialogHint, true, "shiftKey");
-    _registerModifier("advModifierGlobal", advModifierLabel, advModifierHint, true, "altKey");
-    _registerModifier("disAdvModifierGlobal", disAdvModifierLabel, disAdvModifierHint, true, "ctrlKey");
-
-    _registerModifier("showRollDialogModifierLocal", showRollDialogLabel, showRollDialogHint, false, "default");
-    _registerModifier("advModifierLocal", advModifierLabel, advModifierHint, false, "default");
-    _registerModifier("disAdvModifierLocal", disAdvModifierLabel, disAdvModifierHint, false, "default");
 }
+
+const makeName = (name, modifier) => `${name} (${modifier})`;
+const makeHint = (hint, modifier) => `${hint} ${modifier}`;